rename smvpydataset module to smvdataset (#744)
